Switch to highly inefficient but simpler implementation
This version requires way more allocations and copying, but allows
us to inspect and modify the rendered subtree, which could be useful
for indenting lists or wrapping text?

At any rate, the HTML documents we're working with are small enough
where this still finishes instantaneously in debug mode, so it should
be fine.